テーブルごとにファイルを作成する
