+++ drizzle.config.ts
@@ -2,7 +2,7 @@ import '@/database/env-config';
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './src/database/schema.ts',
+  schema: './src/database/schema',
   out: './migrations',
   dialect: 'postgresql',
   dbCredentials: {
